+++ Natery.MultiPhaseProcessor/Natery.MultiPhaseProcessor.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28621.142
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MultiPhaseProcessor", "MultiPhaseProcessor\MultiPhaseProcessor.csproj", "{3D1ABB7C-FA1D-4F4E-982D-2A610C826D56}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Natery.MultiPhaseProcessor", "Natery.MultiPhaseProcessor\Natery.MultiPhaseProcessor.csproj", "{3D1ABB7C-FA1D-4F4E-982D-2A610C826D56}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MultiPhaseProcessor.Test", "MultiPhaseProcessor.Test\MultiPhaseProcessor.Test.csproj", "{241D41E2-A14C-4F67-BBCB-BFE41A078E6C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Natery.MultiPhaseProcessor.Test", "Natery.MultiPhaseProcessor.Test\Natery.MultiPhaseProcessor.Test.csproj", "{241D41E2-A14C-4F67-BBCB-BFE41A078E6C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{67B34EBD-43DA-4F5E-BBDD-1F234576AAD8}"
 	ProjectSection(SolutionItems) = preProject
