+++ Natery.MultiPhaseProcessor/Natery.MultiPhaseProcessor/Processee.cs
@@ -2,63 +2,59 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace MultiPhaseProcessor
+namespace Natery.MultiPhaseProcessor
 {
-    public class HeadProcessee<TInput, TOutput> : IHeadProcessee<TInput>, IProcessee<TInput, TOutput>
+    public class Processee<TInput, TOutput> : IProcessee<TInput, TOutput>
     {
         internal IProcessee<TOutput> _next;
         internal Queue<TInput> _queue;
+        internal bool _moreWorkToAdd;
         internal Func<TInput, Task<TOutput>> _action;
 
-        public HeadProcessee(Func<TInput, Task<TOutput>> action)
+        public Processee(Func<TInput, Task<TOutput>> action)
         {
             _queue = new Queue<TInput>();
+            _moreWorkToAdd = true;
             _action = action;
         }
 
-        public void AddData(TInput data)
-        {
-            AddWorkItem(data);
-        }
-
-        public async Task BeginProcessingAsync()
+        async Task IProcessee<TInput>.BeginProcessingAsync()
         {
             var nextProcessing = _next.BeginProcessingAsync();
-
+            
             var currentProcessing = Executor();
-
+            
             await Task.WhenAll(nextProcessing, currentProcessing);
         }
 
         private async Task Executor()
         {
-            while (_queue.Count > 0)
+            while (_queue.Count > 0 || _moreWorkToAdd)
             {
-                var output = await _action(_queue.Dequeue());
-                _next.AddWorkItem(output);
+                if (_queue.Count > 0)
+                {
+                    var output = await _action(_queue.Dequeue());
+                    _next.AddWorkItem(output);
+                }
+                else
+                    await Task.Delay(100);
             }
             _next.NoMoreWorkToAdd();
         }
 
-        public void AddWorkItem(TInput workItem)
+        void IProcessee<TInput>.AddWorkItem(TInput workItem)
         {
             _queue.Enqueue(workItem);
         }
 
         public void NoMoreWorkToAdd()
         {
-            throw new NotImplementedException();
+            _moreWorkToAdd = false;
         }
 
-        public void AddNext(IProcessee processee)
+        void IProcessee.AddNext(IProcessee processee)
         {
             _next = (IProcessee<TOutput>)processee;
         }
-    }
-
-    public interface IHeadProcessee<TInput> : IProcessee
-    {
-        void AddWorkItem(TInput workItem);
-        Task BeginProcessingAsync();
     }
 }
