+++ Natery.MultiPhaseProcessor/Natery.MultiPhaseProcessor/TailProcessee.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Threading;
 using System.Threading.Tasks;
 
-namespace MultiPhaseProcessor
+namespace Natery.MultiPhaseProcessor
 {
     public class TailProcessee<TInput> : ITailProcessee, IProcessee<TInput>
     {
